requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.50.0
notion-client==2.2.1
python-dotenv==1.0.0
pydub==0.25.1
colorama==0.4.6
gunicorn==21.2.0

